+++ R/clock_model_to_xml_operators.R
@@ -5,15 +5,8 @@
 #' @author Richèl J.C. Bilderbeek
 #' @export
 clock_model_to_xml_operators <- function(
-  inference_model,
-  clock_model = "deprecated",
-  mrca_priors = "deprecated",
-  tipdates_filename = "deprecated"
+  inference_model
 ) {
-  testthat::expect_equal(clock_model, "deprecated")
-  testthat::expect_equal(mrca_priors, "deprecated")
-  testthat::expect_equal(tipdates_filename, "deprecated")
-
   # Don't be smart yet
   clock_model <- inference_model$clock_model
   mrca_priors <- list(inference_model$mrca_prior)