+++ R/clock_models_to_xml_operators.R
@@ -1,9 +1,0 @@
-#' Deprecated
-#' @author Richèl J.C. Bilderbeek
-#' @export
-clock_models_to_xml_operators <- function() {
-  stop(
-    "'clock_models_to_xml_operators' is deprecated, ",
-    "use 'clock_model_to_xml_operators' instead"
-  )
-}