+++ tests/testthat/test-clock_models_to_xml_prior_distr.R
@@ -1,3 +1,0 @@
-test_that("deprecated", {
-  expect_error(clock_models_to_xml_prior_distr(), "deprecated")
-})