+++ tests/testthat/test-clock_models_to_xml_tracelog.R
@@ -1,3 +1,0 @@
-test_that("multiplication deprecated", {
-  expect_error(clock_models_to_xml_tracelog(), "deprecated")
-})