+++ tests/testthat/test-parameter_to_xml_clock_rate.R
@@ -1,3 +1,0 @@
-test_that("deprecated", {
-  expect_error(parameter_to_xml_clock_rate(), "deprecated")
-})