+++ tests/testthat/test-parameter_to_xml_m.R
@@ -1,7 +1,0 @@
-test_that("deprecated", {
-  check_empty_beautier_folder()
-
-  expect_error(parameter_to_xml_m(), "deprecated")
-
-  check_empty_beautier_folder()
-})